task_2.3.1_v4

+++ src/main/java/org/example/crud/controllers/UsersController.java
@@ -25,7 +25,7 @@ public class UsersController {
     }
 
     @GetMapping("/{id}")
-    public String show(@PathVariable("id") int id, Model model){
+    public String show(@PathVariable("id") long id, Model model){
         model.addAttribute("user", userService.getUserById(id));
         return "users/details";
     }
@@ -42,19 +42,19 @@ public class UsersController {
     }
 
     @GetMapping("/{id}/edit")
-    public String edit(Model model, @PathVariable("id") int id){
+    public String edit(Model model, @PathVariable("id") long id){
         model.addAttribute("user", userService.getUserById(id));
         return "users/edit";
     }
 
     @PatchMapping("/{id}")
-    public String update(@PathVariable("id") int id, @ModelAttribute("user") User user){
+    public String update(@PathVariable("id") long id, @ModelAttribute("user") User user){
         userService.update(id, user);
         return "redirect:/users";
     }
 
     @DeleteMapping("/{id}")
-    public String delete(@PathVariable("id") int id){
+    public String delete(@PathVariable("id") long id){
         userService.delete(id);
         return "redirect:/users";
     }
